tried lua method and reverted
